Removed the function from the lib file and created a separate file for the function
